remove old stable structures
