+++ canister/src/lib.rs
@@ -5,7 +5,6 @@ mod blocktree;
 mod guard;
 mod heartbeat;
 pub mod memory;
-mod memory_new;
 mod metrics;
 mod multi_iter;
 pub mod runtime;
